add core space rpc builder
