+++ crates/rpc/rpc-builder/src/cfx/constants.rs
@@ -1,0 +1,2 @@
+pub const DEFAULT_HTTP_PORT: u16 = 12537;
+pub const DEFAULT_WS_PORT: u16 = 12538;
