--- crates/rpc/rpc-builder/src/constants.rs
@@ -1,2 +1,0 @@
-pub const DEFAULT_HTTP_PORT: u16 = 18545;
-pub const DEFAULT_WS_PORT: u16 = 18546;
