+++ crates/rpc/rpc-builder/src/eth/constants.rs
@@ -1,0 +1,2 @@
+pub const DEFAULT_HTTP_PORT: u16 = 8545;
+pub const DEFAULT_WS_PORT: u16 = 8546;
