+++ crates/rpc/rpc-builder/src/eth/error.rs
@@ -26,7 +26,7 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
 // IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
-use crate::EthRpcModule;
+use super::EthRpcModule;
 use std::{
     collections::HashSet,
     io::{self, ErrorKind},
